+++ sub_accounts/sub_account_update_integration_test.go
@@ -1,9 +1,10 @@
 package sub_accounts_test
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
 	"time"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestIntegrationSubAccount_UpdateSubAccount(t *testing.T) {
@@ -69,6 +70,35 @@ func TestIntegrationSubAccount_UpdateSubAccountWarmRetention(t *testing.T) {
 			getSubAccount, err = underTest.GetSubAccount(int64(subAccount.AccountId))
 			assert.NoError(t, err)
 			assert.Equal(t, "test_after_update", getSubAccount.AccountName)
+		}
+	}
+}
+
+func TestIntegrationSubAccount_UpdateSubAccountSoftGBLimit(t *testing.T) {
+	underTest, email, err := setupSubAccountConsumptionIntegrationTest()
+
+	if assert.NoError(t, err) {
+		createSubAccount := getCreateOrUpdateSubAccount(email)
+
+		subAccount, err := underTest.CreateSubAccount(createSubAccount)
+		if assert.NoError(t, err) && assert.NotNil(t, subAccount) {
+			defer underTest.DeleteSubAccount(int64(subAccount.AccountId))
+			time.Sleep(time.Second * 2)
+			getSubAccount, err := underTest.GetSubAccount(int64(subAccount.AccountId))
+			assert.NoError(t, err)
+			assert.Equal(t, "tf_client_test", getSubAccount.AccountName)
+			createSubAccount.AccountName = "test_after_update"
+			softGBLimit := float32(1)
+			createSubAccount.SoftLimitGB = &softGBLimit
+			time.Sleep(time.Second * 2)
+			err = underTest.UpdateSubAccount(int64(subAccount.AccountId), createSubAccount)
+			assert.NoError(t, err)
+			// verify that the update was made
+			time.Sleep(time.Second * 2)
+			getSubAccount, err = underTest.GetSubAccount(int64(subAccount.AccountId))
+			assert.NoError(t, err)
+			assert.Equal(t, "test_after_update", getSubAccount.AccountName)
+			assert.Equal(t, softGBLimit, getSubAccount.SoftLimitGB)
 		}
 	}
 }