+++ sub_accounts/sub_account_update_test.go
@@ -3,12 +3,13 @@ package sub_accounts_test
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/logzio/logzio_terraform_client/sub_accounts"
-	"github.com/stretchr/testify/assert"
 	"io"
 	"net/http"
 	"strconv"
 	"testing"
+
+	"github.com/logzio/logzio_terraform_client/sub_accounts"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestSubAccount_UpdateValidSubAccount(t *testing.T) {
@@ -80,6 +81,32 @@ func TestSubAccount_UpdateSubAccountWithWarmTier(t *testing.T) {
 	updateSubAccount := getCreateOrUpdateSubAccount("test@user.test")
 	warmRetention := int32(5)
 	updateSubAccount.SnapSearchRetentionDays = &warmRetention
+	err = underTest.UpdateSubAccount(subAccountId, updateSubAccount)
+	assert.NoError(t, err)
+}
+
+func TestSubAccount_UpdateSubAccountWithSoftGBLimit(t *testing.T) {
+	underTest, err, teardown := setupSubAccountsTest()
+	assert.NoError(t, err)
+	defer teardown()
+
+	subAccountId := int64(1234567)
+
+	mux.HandleFunc("/v1/account-management/time-based-accounts/", func(w http.ResponseWriter, r *http.Request) {
+		assert.Equal(t, http.MethodPut, r.Method)
+		assert.Contains(t, r.URL.String(), strconv.FormatInt(int64(subAccountId), 10))
+		jsonBytes, _ := io.ReadAll(r.Body)
+		var target sub_accounts.CreateOrUpdateSubAccount
+		err = json.Unmarshal(jsonBytes, &target)
+		assert.NoError(t, err)
+		assert.NotNil(t, target)
+		w.WriteHeader(http.StatusNoContent)
+	})
+
+	updateSubAccount := getCreateOrUpdateSubAccount("test@user.test")
+	softGBLimit := float32(1)
+	updateSubAccount.SoftLimitGB = &softGBLimit
+
 	err = underTest.UpdateSubAccount(subAccountId, updateSubAccount)
 	assert.NoError(t, err)
 }